Move MessageNotSent to wire-api
